+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 - Add support for platform version `231.*`
 - Automatically preview material at caret
+- Add support for scalar map channel option (`-imfchan`) in 3D preview (only RGB)
 - Settings for MTL file editor layout separate from OBJ file editor layout
 - Setting for generating mipmaps for textures (on/off)
 - Setting for selecting face culling